// src/components/organisms/stepper-pulsar/core/stepper-actions.enum.ts
/**
 * Stepper action types
 */
export enum StepperActionsKind {
  INIT = 'INIT',
  RESET = 'RESET',
  ADD_STEP = 'ADD_STEP',
  SET_FIELDS = 'SET_FIELDS',
  GO_TO_STEP = 'GO_TO_STEP',
  COMPUTE_SIBLING = 'COMPUTE_SIBLING',
  COMPUTE_VALIDATION = 'COMPUTE_VALIDATION',
  SET_DATA = 'SET_DATA',
  SET_SUBMITTED_TIMES = 'SET_SUBMITTED_TIMES',
  SET_VISIBLE = 'SET_VISIBLE',
  NAVIGATION_REQUEST = 'NAVIGATION_REQUEST',
  ADD_ERROR = 'ADD_ERROR',
  SUBMIT_REQUEST = 'SUBMIT_REQUEST',
  REMOVE_ERROR = 'REMOVE_ERROR',
  CLEAR_ALL_ERRORS = 'CLEAR_ALL_ERRORS',
  SET_FIELD_DEFAULT_VALUE = 'SET_FIELD_DEFAULT_VALUE',
  SET_STEPPER_DEFAULT_VALUE = 'SET_STEPPER_DEFAULT_VALUE',
}